[dhcp] Correctly respond to timeouts
The DHCP client should always respond to renew, rebind and lease
expiration timers as soon as they fire.

When acquisition timeout is larger than the time gap between now and the
next timeout (e.g. acquisition timeout is 60s, and the lease expires in
5s), the client gets stuck in the current state, keeps retransmitting
requests, when it should have transitioned to the next state. This is
even worse when a very large acquisition timeout is in place, which can
cause the client to skip REBIND altogether.

Test:
fx test netstack_dhcp_gotests

Fixed: 128, 40221

Change-Id: I7dca154a65f97578c04aaf147a29d77d7911dd5d