[testsharder] Make maxShardsPerEnvironment configurable
It was previously a hardcoded constant, but is now controlled via the
-max-shards-per-env flag (which defaults to the old constant values of
8). This will let us configure the max shard count on a per-builder
basis, since our capacity constraints afford more flexibility with some
builders (e.g. that run on x64 QEMU and/or Linux) than others (e.g. that
run on Astro and/or arm64 QEMU).

I'll enable this in follow-up CLs by adding a `max_shards_per_env` spec
field and having the recipes pass `-max-shards-per-env` if the spec
field is set.

Bug: 48650
Change-Id: Ic3f11511f00d20d5c597e679ba6f5e9c636a7ed6