[inspect] Add inspect start timestamp in C++
The health node now reports a timestamp at which the health node was
initialized.

This was added to the deprecated C++ code, but not to the non-deprecated
C++ code.  Adding now.  See the companion change fxr/369233.  In
contrast to the deprecated version, this one does not use abs_clock,
since it is not allowed in zircon code. Instead a clock closure is used.

Bug: 47259, 48335, 44683
Change-Id: I08a0a3ea0cde7235479793467482dc391a5573ba