[owners] Add src/lib/zircon/rust/OWNERS
Prior to this change, ownership of this directory bubbled up to global
owners.
The contents of this file are the intersection of global OWNERS and
people who authored changes under this directory.
This change is a no-op from an ownership perspective. It does improve
owners suggestions for changes here, and creates an affordance for
adding new non-global owners.

Change-Id: I0afa6492ace3e6b2f562b20a3920ef630f61719c