[audio_core] Fix frame alignment in EffectsStage
EffectStage does some alignment of the block requests to support the
constraints that the effect implementations have. The logic to honor
this alignment was incorrectly truncating the frame number to 32-bits.
When the mixer needs to pull data from this stage that is above
UINT32_MAX, we were returning a frame number with the upper 32 bits
masked out. This resulted in the mixer discarding that buffer and
continuously asking for more data, essentially trying to process its way
back through the first 32-bits of frames.

Fix the alignment mask to support 64 bit frame numbers and remove some
ill-advised usages of 'auto' in this portion of code.

Test: fx run-test audio_core_unittests
      New unit test fails without the patch to effects_stage.cc
Fixed: 48641
Change-Id: I82ef682c20193e9085c184786fb1d80211d5488f