[debugger] Show a warning for no attached job.
Attaching to a process by name normally works automatically because zxdb
auto-attaches to the root job. However, if this fails (either
permissions or there is already a debug_agent running that grabbed it),
there is no notification and attaching to processes by name mysteriously
fails.

This adds a warning when attaching by name and there are no currently
attached jobs.

Bug: 35066
Change-Id: Ie9893b78b350983270b0f7f1df78031d3ab46105