[dev_finder] List command does not return error when no devices are found
Finding no devices that match the filters provided is not an error
so the dev_finder command should exit successfully.
This is needed by the Flutter tool which otherwise has to parse
the error message and ignore it when no devices are found.
This change also makes the behavior of dev_finder consistent
with similar tools such as adb which exit successfully when
no matching devices are found.

Bug: 48563
Change-Id: If6f298498a23607fb6ace57e408e566a5a053876