[modular] Fix out-of-order teardown between memfs and ledger.
Memfs for the ledger process was being initialized after, and destroyed
before, the ledger process itself. This caused a race condition that in
most cases resulted in some transient and non-fatal errors in the ledger
process. Occasionally it would result in a fatal crash and interruption
of the normal teardown procedure for Modular.

Fixed: 48634
Test: fx test -o agent_session_restart_test -- --gtest_repeat=1000 --gtest_break_on_failure=1 --gtest_filter=AgentSessionRestartTest.AgentCanRestartSession
Change-Id: I17e90c9aade1f768a55455b908363e0155820008