[policy] Create a appmgr_policy_eng group.
This change adds a appmgr_policy_eng group so that we don't have to have
a bunch of additional definitions in the core configuration. This will
also be used to make it easier to switch between eng, userdebug and
user. All user and userdebug builds are removed in this change.

Bug: 37249
Change-Id: I9999c7cb9c35a0229a794ed4d088108cb6f28e30