[perfcompare] Add "run_local" subcommand for running perf tests locally
See the docs in README.md and in the help strings.

This subcommand will run a Fuchsia performance test and collect the
results in the format accepted by the "compare_perf" subcommand.  It
does the simple (but useful) job of rebooting Fuchsia and running the
test multiple times.

This is intended to be a building block that can be used with
different ways of running tests.  It works with SL4F tests via "fx
run-e2e-tests", but if necessary you can also pass more complex
commands to --iter_cmd like this:

--iter_cmd='fx shell "fuchsia_microbenchmarks -p \
    --runs=10000 --quiet --random-order \
    --filter=\"RoundTrip.*(BasicChannel|ChannelPort|ChannelCall|Futex)\" \
    --out=/tmp/results.fuchsiaperf.json" && \
  fx scp "[$(fx get-device-addr)]":/tmp/results.fuchsiaperf.json \
    tmp_results.fuchsiaperf.json' \
--iter_file=tmp_results.fuchsiaperf.json

Bug: 23104
Test: unit test added to "python perfcompare_test.py" (run by bots on CQ)
 * tested manually against tspec_example_test using steps in README.md
Change-Id: Ibb584de37ec00fd5ad0542831a92b993bb364f32