[debugger] Display variables for every stop.
Add a "display" setting that is a list of expressions to print for every
stop. This can use the existing setting system to create global,
process, and thread-specific lists of variables, and to manage the
contents.

Adds a "display" command to append a new expression to this setting.
This is more natural since it doesn't require quoting, and matches other
console debuggers.

Adds a way to clear a setting to the default value.

Bug: 42233

Change-Id: I25a231df1bc1c3f7b82b9ac77c43f3650a50695f