[appmgr] Make ambient-exec allowlist required
Before this patch, we would only enforce the allowlist for ambient
replace-as-exec if we found such a list on the filesystem.  But that
means that a filesystem failure at the right moment could cause us to
fail open.  We'd prefer to fail closed.

To that end, this patch removes support for optional allowlists in
appmgr.  All the allowlists besides the ambient replace-as-exec
allowlist are already mandatory. We added support for wildcards in
allowlists and a wildcard allowlist intended for use in core in
https://fuchsia-review.googlesource.com/c/fuchsia/+/372882

Fixed: 47836
Change-Id: Iac8686906ae4c68c54e3aa839db293ec755ac6f5