[gn sdk] Prevent gsutil from interactively prompting for updates
gsutil will periodically check for updates and then prompt the user
if they would like to update. We do not want to expose this or cause
automated testing to hang, so we need to prevent this. The update
is prevented if any of stdin, stdout, or stderr are not a TTY, and
redirecting /dev/null in means we do not have to change any arguments.

Fixed: 44472
Change-Id: I337259071dd01b03caf2b31445f76bdafd59f5f5