[gn sdk] Changing to use ssh-keygen for keys
This replaces the assumption that there are keys available in the
running ssh agent. This also makes it easier to copy keys from
the Fuchsia source tree project and projects using the the GN SDK.

This also enables automated tests to run which are run in environments
that have no ssh agent running and configured.

The keys are stored in the Fuchsia SDK data directory in a directory named .ssh.
This is the same structure as used "in-tree" for Fuchsia development. You can copy the
keys from the other directory to make the keys identical, allowing SSH access using both
SDK commands and in-tree "fx" commands.

The authorized key file used for paving is in .ssh/authorized_keys.
The private key used when ssh'ing to the device is in .ssh/pkey.

The second part of is the sshconfig file used by the SDK when using SSH.
This is stored in the Fuchsia SDK data directory named sshconfig.
This script checks for the private key file being referenced in the sshconfig, if it
is not present, the sshconfig file is regenerated.
You can customize the sshconfig and it will not be overwritten unless the full path to the
private key is not present. If you want to remove it as part of customization, you can comment out
the IdentityFile line with the `#` character.

Change-Id: I045ed3e616675c3bed015bb10bbf7e0e005081f5