[gn-sdk]Add far static analysis test
This CL adds a far test that ensures that some libraries that are
depended on by each component contained in a far archive
(e.g. ld.so.1) are also caontained in that far archive. This test
also ensures a component manifest is present in the far archive.

The purpose of this test is to increase the test coverage of the
GN SDKs generation to catch issues before they are propogated to
downstream customers (e.g. Chromium).

Change-Id: I475a3c45825094b9cc51cb2bc118730426beb40e