[PowerManager] Node setup using config file
The PowerManager node configuration is currently initialized at runtime
by querying the current board config and creating the appropriate nodes
with product-specific configuration accordingly. The better approach is
to have product-specific node and configuration data live in a separate
configuration file that the PowerManager can consult and use to
initialize itself. This will allow us to move forward with PowerManager
bringup on future products beyond Astro.

Change-Id: I13692d01723c1cbc9672c5f36f56c9911b2cedd4