[kernel][object] Remove RootJobObserver dependency on Dpc
Previously, RootJobObserver fired off a Dpc to itself from inside an
OnStateChange method.  Although the OnStateChange method is called
under the root job Dispatcher's lock there is no need to use a Dpc as
the method does not call anything that (transitivity) attempts to
acquire the Dispatcher's lock.

We remove the dependeny on Dpc because Dpc is difficult to use
correctly and is not necessary here.

Tested by building with lockdep and killing off the root job's
children.

Bug: 48647
Change-Id: I5592fcc26eb6cdf347e0b364a655442ee019c347