[debugger] Lookup escaped identifiers.
Previously the parser put the "$(...)" (used to escape identifers) into
the identifier components despite the header file for the "escaped
identifier" enum saying this didn't happen.

The result is that when code went to look up the symbol with this name,
it would get the literal "$" and parens and lookup would fail.

This patch strips the "$(" and ")" from escaped identifiers when putting
the strings in the contents. It adds code to put these back in for
printing to the screen when necessary for parsing.

Change-Id: Ie8f56c4e3c680b1b038b0ed7e59240e710c6975e