[kernel][vm] VmPageList callbacks receive page ownership
This changes all VmPageList functions to rely on callbacks taking
ownership of pages such that the VmPageList never needs to put pages
into freed lists.

The motivation for not having the VmPageList put freed pages into lists
is that the PageQueues have ownersip of the queue node whilst a page is
in a queue. If the VmPageList manipulates the queue node to place pages
into free list, then it forces callbacks to always immediately ensure
that the PageQueue does not own the queue node. This is too rigid a
requirement and imposes performance limitations in future implementation
work.

As any VmPageList method that takes a list_node_t with the intent to add
pages to it will eventually strictly not be used by the VM code they are
all removed. This ensures that the unittests are forced to use the
versions that the rest of the VM code will actually use.

Test: No functionality changes. Refactor is covered by existing tests
in vm_unittests and the utest/core/vmo tests

Change-Id: If1450c847f49cf2656ec1d6a474d180e8ee93d75