[audio][visalia] Add DC removal filter
This fixes the huge DC observed by shyamsundarp@ in
fxb/45385. Note that hollande@ is investigating the root
cause of the current CIC filter's DC component, we may
either improve or replace the whole filter based on that
investigation. In any case though, the current implementation
must not have such amount of DC exported to clients of the driver.
This CL adds DC removal by calculating the DC in the data
using a simple low pass filter.

Fixed: 45385

Test: runtests -t audio-dsp-test.
Test: record audio in visalia and check for DC offset.

Change-Id: I308cc6ce86e9f495d49dfca89df20983800b19c5