[stash] Introduce secondary store
Add --secondary_store flag to serve a fuchsia.stash.Store2 that is a
clone of Store in all but name.

A follow-up change will modify one of the existing Store clients to use
Store2, and modify sysmgr to serve a Store2 instance in the sys
environment.

This is a minimal change to solve fxb/48802.
Per go/fxb-47602-research we don't expect to need more than this.
Therefore I'm ok with a very non-extensible solution to this problem.
The next step, if there is one, is to use CFv2 features such as
directory capabilities or hashed monikers to introduce more isolation
domains, not to further generalize this solution.

Bug: 48802
Change-Id: I500b8104653fb29f17b7435ccbfeb442bcbd8f93