[dhcp] Deflake client state transition test
This test is flaky because of a race condition where the test checks
client state right after address is acquired, and it is possible client
info (containing the state) is not updated yet, causing the test to
flake. It is OK to remove client state checks because acquired
addresses are still checked, which confirms state transition.

Test:
fx test -o netstack_dhcp_gotests -- \
  -test.run TestStateTransition/LeaseExpire \
  -test.count 100

Fixed: 48962

Change-Id: I4c731cc343d0a36ccd6817cf6aa1b230d1a664ba