[feedback] CrashpadAgent sets Crash guid
We want to use the DeviceId as the guid for all crash reports, if one
exists. We asynchronously fetch the DeviceId and cache the value
returned. Multiple requests may be made to the wrapper concurrently,
but only one call is made per connection to the server.

TESTED=`fx test crashpad_agent_tests`
TESTED=`uploaded crash reports with and without the change:
> with: 4759fb08b054b0ec
> without: 4c421502f0415991
and each has a different ClientID.`

Fixed: 42590

Change-Id: I8475cae428346a99d37a211f0b3896f74d05309e