[i18n] Remove l10n items from translation console
These files are unused in the current build process,
so removing them to reduce clutter.

The removed files are remnants from the files generated by
Google's translation process.  At the moment we're using
the ARB-based process for localization, so these files are
not fully compatible. If we decide to switch to Google's
proces, we can always reinstate these files.

Fixed: 49058
Change-Id: I52e3cbfca4a3664d8d9396a81a2d484346e037fb