[elflib] fix AArch64 plt discovery
AArch64 uses two instructions to compute an offset in a table. The first
one is adrp which computes the high bits of the offset and puts it in a
register. The second one is ldr which loads a value in the table with the
previously computed register and a 12 bit offset.

The previous implementation only used ldr to infer the index in the plt.
This was not correct because, for big plts, the adr instruction can
change from one entry to the other. Most of all, the plt entries don't
have the same order as the plt symbols.

This patch defines a index which is incremented to access the plt
symbols.

Bug: 40897
Change-Id: I226ec81375b214a930809b41c8ec5ce21b29691f