[pkgfs] Support restrictions on execute rights
This change adds a new mode to pkgfs that, if enabled, restricts
executability rights to packages in either the static index or an
allowlist. This mode can be statically disabled based on the presence
of a data file in the system_image package; if that file is not present,
the feature can be enabled with a new pkgfs command line flag.

Bug: 48568

Change-Id: I5ff2a5f0ad85fc598aa8aaa640e6103a37e305de