[fidlc] Removing FragileBase attribute
Until we introduced 64-bit method ordinals, there was a high probability
of breakage at a distance when methods were added in a composed protocol
(a "parent") which would clash with a method defined in protocol
composing it (a "child"). To make library authors aware of this problem,
we forced the use of the `[FragileBase]` attribute.

This is no longer needed, and this requirement was dropped in FTP-029.

PS: The real story is abarth@ wanted to make sure pascallouis@ did not
forget about this breakage at a distance problem, and asked in FTP-020
to keep this relatively unelegant attribute as an eyesore reminder. And
voila, work done.

Test: fidlc tests
Change-Id: Ic6393cab1705b42127e5aa86759adf596a8f3514