[gn sdk] Enabling fidl metadata for fidlcat
This adds the ability to generate a file containing
the paths to all the fidl intermediate representations.
This file is useful for passing to fidlcat using the
--fidl-ir-path command line argument.

To create the file typically named "all_fidl_json.txt" in the
root of of the output directory, add a target to your top level
BUILD.gn file such as:

generated_file("all_fidl_json") {
  testonly = true
  deps = [ "//:default" ]

  outputs = [ "$root_build_dir/all_fidl_json.txt" ]
  data_keys = [ "fidl_json" ]
}

Replacing "//:default" with the dependencies that should be traversed to
collect the fidl files of interest.

Change-Id: I48ea12f0195b9d6509e9d2dc82ab3d7c255fc7c0