[fx][sl4f][wfh] Support custom HTTP and SSH ports in SL4F
Change SL4F client:
- custom HTTP port (SL4F_HTTP_PORT env variable or constructor argument)
- custom SSH port (FUCHSIA_SSH_PORT env variable)
- change the name of the env variable that defines the device address to
FUCHSIA_DEVICE_ADDR to avoid a potential misunderstanding that IPv6
would not be supported with the previous name (FUCHSIA_IPV4_ADDR).
The previous name is still supported for legacy clients, and if both are
defined, FUCHSIA_DEVICE_ADDR takes precedence.

Change contrib/run-e2e-tests:
- export FUCHSIA_SSH_PORT if vars->get-device-ssh-port is not empty
- export FUCHSIA_DEVICE_ADDR with vars->get-fuchsia-device-addr. The
legacy env variable FUCHSIA_IPV4_ADDR is still exported for legacy uses.

Existing behavior changes in clients using the Sl4f constructor
directly. Previously, a target address could be specified in the
constructor with the 'address:port' syntax and that was used in some
tests. However, if used in real code, it would cause issues because
a few e2e tests had assumptions that 'sl4f.target' was only an address,
without port - they append ':port' for their own debug agents.
This CL enforces it by throwing an exception if a ":port" suffix is
found in the "address".

Bug: b/152335383
Change-Id: Ia7e7f37e358032565e406cd4bdafd50a91d6cffd