[pkg][repo] Expose utility to get root keys from a repo
A functional change to expose an existing utility to derive the root
keys of a package repository. This will be useful for the current caller
to do this, as it must deal separately with the same address scoped
differently for the host and target (i.e., it must get the root keys
from the former, but pass on the latter to a device).

Bug: 10427
Change-Id: Ic6217485f346dcccec3c37f193a4acdb4d00ea67