[zx] replace_as_executable: take resource& not handle&
If you have a zx::resource, you should be able to borrow it for
zx::vmo's replace_as_executable().  However, with the declarations as
written, that's not possible -- the function takes a zx::handle&, which
you can't safely get an instance of from a valid zx::resource without
releasing the underlying handle or destroying the zx::resource, which
rather defeats the point.

Instead, this API should probably always have taken a zx::resource&.
We failed to notice this because we'd always used the invalid handle
in all calls to date, and are only now getting around to trying to use
the actual capabilities as intended.

To allow the caller in chromium to make a soft transition, we add an
additional implementation that takes a zx::resource& and migrate all
in-tree callers to this pattern.  We'll move chromium to the new
implementation once available, and will remove the old implementation
that takes a zx::handle& once all out-of-tree callers have migrated off.

There were two callers in hermetic-compute that used the implicit {},
which this patch made ambiguous, but they can be converted to use
explicit zx::resource() constructors.

Bug: 45994
Change-Id: Ibb82422cdc478b1bdbedc139e2be135355c9e0cc