[fidl][rust] Optimize Encoder::reserve_out_of_line
This CL optimizes Encoder::reserve_out_of_line for the case where the
new length is greater, but still within the buffer capacity (no
allocation needed). In this case, it now uses an unsafe block to change
the buffer length and memset the new elements to zero.

Profiling revealed the old implementation (Vec::resize) to take ~97% of
FIDL encoding time for a vector<uint8> of length 8192, in non-release
mode. It got optimized in release mode, but we still chose to make this
change because (1) we prefer more acceptable performance even in
non-release mode, and (2) in release mode, we feel it is too brittle to
rely on the compiler optimizing the loop in Vec::extend_with in exactly
the way we need it.

Test: CQ
Bug: 45055
Change-Id: I06d17c6f35ded25de6a5a6da0e5f392d01b7d341