[system-tests] Fix run-test with multiple arguments
The `tmux send-keys arg1 arg2 ...` will actually remove spaces between
each argument in order to interpret arguments `C-m` as a carriage return.
In order to preserve spaces, we need to quote the arguments. Previously,
when we used "$@", it actually gets expanded into
`tmux send-keys "arg1" "arg2" ...`, so it'll also drop the spaces. In order
to make this work, we need to instead explicitly quote the arguments
with `printf "%q " "$@"`, and then wrap that string in quotes in order to
get the correct behavior.

Test: Verified that the `run-test ... system_tests_upgrade --ssh ...`
actually works.

Change-Id: I7a5841cb8e38681036a7679beb59a97c823595ee