[gn sdk] Adding fserve-remote.sh
This command allows the starting of fserve.sh from a remote computer and
sets up tunneling via SSH between the host and the remote so the device
connected to the remote computer can be accessed via the tools and
package requests from the device are forwarded to the host computer.

This is modeled after the fx serve-remote command use in the fuchsia
tree.

Change-Id: I997190b36e59a9cde60714c5669fe78364aada12