[hwreg] Add PIO and mock support
This renames hwreg::RegisterIo to hwreg::RegisterMmio and gives it a
sister type hwreg::RegisterPio.  PIO is supported either as direct
in/out instructions (x86 only) or as a scaled mapping to MMIO.  It
also adds hwreg::Mock, which can be used in place of
hwreg::Register*io for testing.  In general it's now supported to
use a std::variant<...> type of several alternative types that all
support the common hwreg templated Read/Write API.

Change-Id: I8b2254840367c71cc139b29c0342fdf0cd544679