[emu-remote] add script for running emulator remotely
This can be used to run the emulator remotely and stream
the output to a local browser. Automatic GPU mode
selection is now done by the emulator and X server access
failure detection has been improved to better support
this use-case.

Basic example:
$ fx emu-remote HOST

With networking:
$ fx emu-remote HOST -- -N

With networking and GPU acceleration:
$ fx emu-remote --display :0 HOST -- -N

Change-Id: I7a881e2a37cda1183722ae3d3a36630e6313048e