[graphics][compute] Implement swapchain staging.
This CL allows the programs using the graphics compute test framework,
like the spinel and Mold demo programs, to work properly on laptops with
Intel integrated GPUs, whose swapchain implementation does not support
VK_IMAGE_USAGE_STORAGE_BIT, as required by Spinel.

This is achieved by implementing "staging", where an intermediate
target image is provided to clients to render to, and to ensure that
vk_swapchain_present_image() will transfer the target image's content
to the real swapchain.

This works because the target image is created with a format different
from the swapchain, but with an image view that swaps the R and B
channels automatically.

There is a non-trivial performance impact (i.e. one full surface copy
per swapchain presentation event), but this is the price to pay to be
able to run the demos on corp Linux laptops.

Bug: 47665
Test: out/default/host_x64/spinel-svg-demo SVG_INPUT_FILE
      will display an input SVG document with proper colors.
Change-Id: I4a5ca6bb3099680755e243526887f2fd1b0f47cd