[rust][fidl] Use Vec::extend_from_slice for copy encoding
This CL makes the Rust bindings use Vec::extend_from_slice when encoding
vectors of primitives and strings. This is efficient because (1) it
compiles to a memcpy, even in debug mode, and (2) it writes directly to
uninitialized memory instead of zeroing it out first.

Recently there has been an effort to optimize vector<uint8> encoding
which resulted in the following changes:

1. https://fuchsia-review.googlesource.com/c/fuchsia/+/368303
   This special-cased vectors of primitives to use the type &[T] and
   encode with std::ptr::copy_nonoverlapping.

2. https://fuchsia-review.googlesource.com/c/fuchsia/+/374968 This
   optimized Encoder::reserve_out_of_line to use Vec::set_len and
   std::ptr::write_bytes instead of calling Vec::resize, which was
   taking a lot of time in traces.

Both these changes required unsafe blocks. This CL is the result of
realizing that we can achieve the same performance gains as (1) and (2)
in less code and without using unsafe blocks, simply by using
Vec::extend_from_slice. In fact it is even more efficient, since it
_eliminates_ the zeroing pass whereas (2) was just making it faster.

Nonetheless, those earlier CLs were not a waste. The bulk of (1) was the
type change, not the memcpy function. And the optimization from (2) is
still beneficial for types not on the fast path, such as large vectors
of non-primitive types.

Test: fx test rust_fidl_tests
Change-Id: I0beca47aeee0864608122116c5d334783fe541f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375649
Reviewed-by: Benjamin Brittain <bwb@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>