[fx debug] support wfh flows and ipv4
The change appears larger because I folded the library file into the
script, it was sharing a lot of global state with the script that was
hurting readability. As the script is now shorter overall, there is
less need for the extra file.

The two helper functions are now implemented to take parameterized
inputs as opposed to depending on global state and use more fx
built-ins rather than self implementing various functionality. Most
significantly they no longer deal in target addresses, which
simplifies the migration to ipv4 support.

The subprocess is no longer tracked by pid, as the zxdb client will
request it to exit anyway increasing the possibility that we're
racing, instead we now just wait on all subprocesses, which covers
the case.

Bug: 49094
Change-Id: Iefa8b24eb49ce0113b80cf7a6eb9a79e5f5cb4a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375757
Testability-Review: Brett Wilson <brettw@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>