[rust][scoped-task] Add --nocapture to scoped-task-test
There seem to be some flakes of this test. Since I've run it many times
it's unlikely that I'll be able to reproduce (though I will try). In the
meantime, I'd like to get the full output with --nocapture so I can
diagnose the flake.

--nocapture means the output from each test is printed directly to the
log. --test-threads=1 causes the tests to be run one at a time, so their
output is not interleaved. --ensure-time is an experimental option that
causes each unit test to fail after a timeout.

Bug: 49047
Change-Id: Id535cca5919c1a8c7219bcb8fe48c7876b57c7c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376106
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>