[x86][hypervisor] Prevent the guest from writing to host CR0
Some bits of CR0 are not restored when exiting the guest, so when the
guest writes to them those values persist for the host. This CL masks
those values and also shadows them so that the guest reads back the same
values that it writes.

BUG: 47699

Change-Id: Id300307087bb129d3f166b31f1f786db77e73e5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375303
Commit-Queue: Alex Legg <alexlegg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Alex Legg <alexlegg@google.com>