[storage] Fix TODOs in minfs/fsck.cc.
When inodes are deleted now, they are updated to indicate that they are
purged.

Also add an oldest_revision field to the superblock so that fsck (and
other code in future maybe) can get an idea of the oldest revision of
code that has touched the volume. fsck can use that to decide whether to
throw errors or warnings.

Change-Id: Id9a1dfcb628897edc7590d7a138a984051209a33
Bug: 48522
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/372960
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Chris Suter <csuter@google.com>