[components] Introduce async EventSource
This CL introduces the ability to request an async EventSource.

This is useful for the Archivist to permit it to observe component
events without blocking component manager or components.

EventSourceSync is a composition of EventSource and `StartComponentTree`
operation. In other words, if EventSource (non-sync) is requested, then
children will not block starting.

However, on the service-side, EventSourceSync is used. On the client
side, either EventSource or EventSourceSync can be used regardless of
which capability is requested. If EventSource is requested,
`StartComponentTree` is ignored.

At the moment, this CL uses a single client library for both testing and
non-testing use cases. A subsequent CL will move testing code into a
separate trait and file.

Change-Id: I621e049f4ad93e878789b4aa3988ea7a67efdd64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374629
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
API-Review: Gary Bressler <geb@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>