[ethernet][ax88179] Save online status and set once started
If the "online" interrupt is serviced before the device is started, the
ifc is never marked as online. To fix this, use the saved online_ status
to set Status() on the ifc_ once it's created.

The fix is the line in
src/connectivity/ethernet/drivers/asix-88179/asix-88179.cc, the rest is
scaffolding to be able to add a test.

New test function driver added in asix-88179-function.cc, test cases for
online bit in asix-88179-test.cc, and fidl to communicate between them
in hooks.fidl (used to force the online status).

Additionally, include a patch from bbosak@ for
UsbPeripheral::UsbDciInterfaceControl(). Without this the added function
driver fails during setup.

Bug: 40786
Test: /system/test/sys/usb-virtual-bus-asix-88179-test --gtest_repeat=10 --gtest_break_on_failure
Change-Id: Ic2a64498c95dc804f275928e8140e78137b49445
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/373849
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>