[memory] Add k command to enter a memory availability state
Add command `k pmm mem_avail_state <state> <nsecs>` to enter the memory
availability state <state> and hold that state for <nsecs>. Can only be
used to transition to a more memory-critical state. Allocates memory to
enter the required state, sleeps for nsecs, then frees the pages that
were allocated.

This is meant as a debug tool to generate memory pressure notifications
and is not necessarily exact. We can race with memory being allocated or
freed on another core while we wait, and the resultant state could be
different from the one that was requested. The command makes sure to
free up all the memory it allocated before exiting.

Debug aid for:
Bug: 45997

Change-Id: I554987c537ebc08f67114da384fb495cf15d5be5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376283
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>