[graphics][aml-gpu] Use per-SoC initial clock
We currently use a starting clock index of 2, which happens to be
middling on all devices. This arbitrary value didn't used to matter
since as soon as thermd started it would pick a better value.

However, thermd is being replaced with power_manager, which doesn't
support GPU throttling. Because of that we want to switch to starting in
the maximum clock frequency on sherlock and astro. We believe we have
enough thermal headroom that the high frequency won't be a problem.

Fixed: 49116
Change-Id: Ia5be8a84dc85082002da8325dc90da99ffb1f3a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376103
Reviewed-by: Craig Stout <cstout@google.com>
Testability-Review: Craig Stout <cstout@google.com>
Commit-Queue: John Bauman <jbauman@google.com>