[botanist] Allow "localhost" placeholder in repo/blob URLs
If provided, "localhost" will be resolved as the remote-scoped local
address. The utility in having this lies in the fact that in the
infrastructure one must construct the botanist commandlne ahead of time,
and we cannot always know the precise local address ahead of time.

Not in use yet - so this change is a no-op - but a recipe change will
follow to start using this codepath.

Bug: 10427
Change-Id: I17d8df674bb8bc920715eb103ffa09fbc879247b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375575
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Testability-Review: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>