[a11y] Adds start event to one finger drag gesture recognizer.
The one finger drag gesture recognizer has support for events to be fired for
drag in progress and drag finished. Adds support for drag started, so that it
can perform an initial setup when necessary, as well as a shutdown at the
complete event.

Change-Id: Ia447e61bbebfbca9725daa8560feac18b7e45041
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375197
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Testability-Review: Alice Neels <neelsa@google.com>