[net-types] Add IpAddressWitnessExt trait
Add the `IpAddressWitnessExt` trait, which adds functionality
to some `Witness<A>` types where `A: IpAddress`. This trait provides
the `with`, `with_v4`, and `with_v6` methods, which are analogous
to the methods on the `IpAddress` trait with the same names,
but preserve the witness.

Concretely, when called on a witness type `W<A>`, the IPv4 closure
takes a `W<Ipv4Addr>` argument, and the IPv6 closure takes a
`W<Ipv6Addr>` argument.

This allows code which operates on witness types of IP addresses
to use `with`, `with_v4`, and `with_v6` just as code which operates
on plain IP addresses is already able to do.

Also, add the `from_witness` convenience constructor to the `Witness`
trait; it accepts a witness type, extracts the inner address, and
then calls the existing `new` constructor.

Test: None

Change-Id: I69e5d849c8d752c0c9982e7b08d9b3da558dc60b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/373693
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>