[ulib][zxc] Introduce Zircon common result type.
Introduce the Zircon common library and the fitx::result type and
family as the first members.

fitx::result is a flexible implementation of the result pattern that
supports user-defined error types and zero or one return values.

zx::status is a utility type that adapts fitx::result to zx_status_t
status codes, making it easier to integrate with existing code.

The library includes a number of helper types and functions for
use with the result/status types, including:
  - ok
  - error
  - failed
  - success
  - as_error
  - make_status
  - error_status

See README.md for a complete explanation of the API.

Bug: 37326
Test: Tests included in this CL pass.
Change-Id: Ief44aed32c254466c9312fdd90166636b225098b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/363193
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Mark Schott <schottm@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Testability-Review: Adam Barth <abarth@google.com>