[audio_core] Device reference clocks
This CL adds a default reference clock to the AudioDevice class.
Initially, this clock is a clone of CLOCK_MONOTONIC. In subsequent
CLs, if the driver indicates that the device is NOT in the same clock
domain as system monotonic, then this clock will diverge from
CLOCK_MONOTONIC as it is fine-tuned to match the device's actual
clock.

Test: 2 new AudioDevice unittests in audio_core_unittests

Change-Id: I439f31dd9f0818d29043a211e0b1170d2ce2c5f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375836
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>