[storage] Don't repair FVM slices until after replaying the journal.
Except for the fields required to be sure of the journal's location,
none of the other fields can be trusted until *after* the journal has
been replayed (because there might be a transaction there that updates
the superblock). That includes the FVM information, so repairing the FVM
slice count needs to be done after replaying the journal.

Bug: 49072
Change-Id: I44547328da6a57ef871f31bcdc6e306edf09ca93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376453
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Chris Suter <csuter@google.com>